socketio-minimal-demo-main/server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-multipart==0.0.6
aiofiles==24.1.0
redis==5.0.1
opencv-python==4.8.1.78
pillow==10.0.1
PyMuPDF==1.23.8
aiohttp==3.9.1